Move pkg/ip to pkg/resolver/external
