+++ pkg/resolver/external/client_test.go
@@ -1,7 +1,7 @@
-package ip_test
+package external_test
 
 import (
-	"cloudflare-ddns/pkg/ip"
+	"cloudflare-ddns/pkg/resolver/external"
 	"cloudflare-ddns/pkg/test"
 	"context"
 	"net/http"
@@ -11,35 +11,35 @@ import (
 func Test_VersionsGet(t *testing.T) {
 	tests := []struct {
 		name     string
-		version  ip.Version
+		version  external.Version
 		want     string
 		wantURL  string
 		hasError bool
 	}{
 		{
 			name:     "ipv4 should return ipv4 address",
-			version:  ip.V4,
+			version:  external.V4,
 			want:     "192.168.0.1",
 			wantURL:  "https://api.ipify.org",
 			hasError: false,
 		},
 		{
 			name:     "ipv6 should return ipv6 address",
-			version:  ip.V6,
+			version:  external.V6,
 			want:     "fe80::bf51:5d53:8f20:9d4",
 			wantURL:  "https://api6.ipify.org",
 			hasError: false,
 		},
 		{
 			name:     "ipv4 should fail and error is expected",
-			version:  ip.V4,
+			version:  external.V4,
 			want:     "",
 			wantURL:  "https://api.ipify.org",
 			hasError: true,
 		},
 		{
 			name:     "ipv6 should fail and error is expected",
-			version:  ip.V6,
+			version:  external.V6,
 			want:     "",
 			wantURL:  "https://api6.ipify.org",
 			hasError: true,
@@ -57,7 +57,7 @@ func Test_VersionsGet(t *testing.T) {
 	for _, tt := range tests {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			client := ip.NewClient(ip.Client(test.NewTestClient(func(r *http.Request) *http.Response {
+			client := external.NewClient(external.Client(test.NewTestClient(func(r *http.Request) *http.Response {
 				if r.URL.String() != tt.wantURL {
 					t.Errorf("wrong URL, want %q, got %q", tt.wantURL, r.URL.String())
 				}
